+++ src/components/button/RelativeButton.tsx
@@ -1,22 +1,6 @@
 import { TouchableOpacity, Text, View } from 'react-native';
 import { styles } from './style.js';
 
-interface Normal {
-  text: string;
-}
-
-export function Normalbutton(props: Normal) {
-  const onClick = () => {
-    console.log('hi');
-  };
-  return (
-    <TouchableOpacity onPress={onClick}>
-      <View style={styles.Normalbutton}>
-        <Text style={styles.buttonfont}>{props.text}</Text>
-      </View>
-    </TouchableOpacity>
-  );
-}
 interface Relative {
   name: string;
   relative: string;
